VID-516 Add logging and reduce cache ttl
Add logging wherever we set fileExists to false to try to trap the condition where the file does actually exist, but memcache has fileExists === false set.

Also reduce the ttl for the cache in the case where fileExists is false so that this problem will be reduced while this problem is investigated.