PLATFORM-280 Replace checked-in PHPUnit 3.7.12 with composer one
No version bump; just make it work with Composer, first. This also DOES
NOT check in the PHPUnit code, because checking in dependencies is a
really bad idea since it suggests that people can modify it at will,
which is exactly what happened last time and should never happen again.